refactor(transform): isolate static evaluation runtime
